consumer/databaseDashboard/requirements.txt
fastapi
uvicorn[standard]
ssi-fc-data  # nếu bạn dùng thư viện này
kafka-python
pandas
sqlalchemy
psycopg2-binary
confluent-kafka